+++ src/main/java/com/bugflix/weblog/config/SecurityConfiguration.java
@@ -1,23 +1,28 @@
 package com.bugflix.weblog.config;
 
-import com.bugflix.weblog.config.security.JwtAccessDeniedHandler;
-import com.bugflix.weblog.config.security.JwtAuthenticationEntryPoint;
 import com.bugflix.weblog.security.JwtAuthenticationFilter;
 import com.bugflix.weblog.security.JwtProvider;
+import jakarta.servlet.ServletException;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.http.SessionCreationPolicy;
+import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.crypto.factory.PasswordEncoderFactories;
 import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.security.web.SecurityFilterChain;
+import org.springframework.security.web.access.AccessDeniedHandler;
 import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
 import org.springframework.web.cors.CorsConfiguration;
 import org.springframework.web.cors.CorsConfigurationSource;
-import org.springframework.web.cors.UrlBasedCorsConfigurationSource;
 
+import java.io.IOException;
 import java.util.List;
 
 
@@ -27,86 +32,90 @@ import java.util.List;
 public class SecurityConfiguration {
 
     private final JwtProvider jwtProvider;
-    private final JwtAuthenticationEntryPoint jwtAuthenticationEntryPoint;
-    private final JwtAccessDeniedHandler jwtAccessDeniedHandler;
 
     private static final String[] PERMIT_TO_USER = {
-            "/api/users/test",
-            "/api/auth/reissue"
     };
 
     private static final String[] SWAGGER_URL_ARRAY = {
             "/v3/api-docs/**",
             "/swagger-ui/**",
-//            "/swagger-resources",
-//            "/swagger-resources/**",
-//            "/configuration/ui",
-//            "/configuration/security",
-//            "/webjars/**",
+            "/swagger-resources",
+            "/swagger-resources/**",
+            "/configuration/ui",
+            "/configuration/security",
+            "/webjars/**",
     };
 
     private static final String[] PERMIT_TO_ALL = {
             "/api/users",
-            "/api/auth/login",
+            "/api/users/login"
     };
 
     @Bean
-    public SecurityFilterChain filterChain(HttpSecurity httpSecurity) throws Exception {
-        httpSecurity
+    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
+        http
                 // ID, Password 문자열을 Base64로 인코딩하여 전달하는 구조
                 .httpBasic().disable()
                 .csrf().disable()
-
                 // CORS 설정
-                .cors().configurationSource(corsConfigurationSource())
-
+                .cors(c -> {
+                            CorsConfigurationSource source = request -> {
+                                // Cors 허용 패턴
+                                CorsConfiguration config = new CorsConfiguration();
+                                config.setAllowedOrigins(
+                                        List.of("*")
+                                );
+                                config.setAllowedMethods(
+                                        List.of("*")
+                                );
+                                return config;
+                            };
+                            c.configurationSource(source);
+                        }
+                )
                 // Spring Security 세션 정책 : 세션을 생성 및 사용하지 않음
+                .sessionManagement().sessionCreationPolicy(SessionCreationPolicy.STATELESS)
                 .and()
-                .sessionManagement()
-                .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-
                 // 조건별로 요청 허용/제한 설정
-                .and()
                 .authorizeRequests()
-
+                // 회원가입과 로그인은 모두 승인
+                // USER Role이 있는 사용자만 허용
                 .requestMatchers(PERMIT_TO_USER).hasRole("USER")
+                // swagger-ui와 관련된 주소 허용
                 .requestMatchers(PERMIT_TO_ALL).permitAll()
                 .requestMatchers(SWAGGER_URL_ARRAY).permitAll()
-                .anyRequest().authenticated() // .anyRequest().denyAll() -> FilterChain.doFilter(...) AccessDeniedException
-
-                // JWT 인증 필터 적용
+                .anyRequest().denyAll()
                 .and()
+                // JWT 인증 필터 적용
                 .addFilterBefore(new JwtAuthenticationFilter(jwtProvider), UsernamePasswordAuthenticationFilter.class)
-
                 // 에러 핸들링
                 .exceptionHandling()
-                .accessDeniedHandler(jwtAccessDeniedHandler)
-                .authenticationEntryPoint(jwtAuthenticationEntryPoint);
+                .accessDeniedHandler(new AccessDeniedHandler() {
+                    @Override
+                    public void handle(HttpServletRequest request, HttpServletResponse response, AccessDeniedException accessDeniedException) throws IOException, ServletException {
+                        // 권한 문제가 발생했을 때 이 부분을 호출한다.
+                        response.setStatus(403);
+                        response.setCharacterEncoding("utf-8");
+                        response.setContentType("text/html; charset=UTF-8");
+                        response.getWriter().write("권한이 없는 사용자입니다.");
+                    }
+                })
+                .authenticationEntryPoint(new AuthenticationEntryPoint() {
+                    @Override
+                    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
+                        // 인증문제가 발생했을 때 이 부분을 호출한다.
+                        response.setStatus(401);
+                        response.setCharacterEncoding("utf-8");
+                        response.setContentType("text/html; charset=UTF-8");
+                        response.getWriter().write("인증되지 않은 사용자입니다.");
+                    }
+                });
 
-        return httpSecurity.build();
+        return http.build();
     }
 
     @Bean
     public PasswordEncoder getPasswordEncoder() {
         return PasswordEncoderFactories.createDelegatingPasswordEncoder();
-    }
-
-
-    /*
-        CorsConfig - https://toycoms.tistory.com/37
-        configuration.addAllowedOrigin("*");
-        configuration.addAllowedHeader("*");
-        configuration.addAllowedMethod("*");
-     */
-    @Bean
-    protected CorsConfigurationSource corsConfigurationSource() {
-        CorsConfiguration configuration = new CorsConfiguration();
-
-        // Cors 허용 패턴
-        configuration.setAllowedOrigins(List.of("*"));
-        configuration.setAllowedMethods(List.of("*"));
-        UrlBasedCorsConfigurationSource source = new UrlBasedCorsConfigurationSource();
-        source.registerCorsConfiguration("/**", configuration);
-        return source;
     }
 }
