--- src/main/java/com/bugflix/weblog/config/security/JwtAuthenticationEntryPoint.java
@@ -1,23 +1,0 @@
-package com.bugflix.weblog.config.security;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-import org.springframework.security.core.AuthenticationException;
-import org.springframework.security.web.AuthenticationEntryPoint;
-import org.springframework.stereotype.Component;
-
-import java.io.IOException;
-
-@Component
-public class JwtAuthenticationEntryPoint implements AuthenticationEntryPoint {
-    @Override
-    public void commence(HttpServletRequest request, HttpServletResponse response, AuthenticationException authException) throws IOException, ServletException {
-        // 인증문제가 발생했을 때 이 부분을 호출한다.
-        response.setStatus(401);
-        response.setCharacterEncoding("utf-8");
-        response.setContentType("text/html; charset=UTF-8");
-        response.getWriter().write("인증되지 않은 사용자입니다.");
-    }
-
-}
